Mosquito.DerivedRules

Description

Derived, useful theorems --- direct consequences of the primitive inference rules defined in the kernel.

Reflexivity

reflexivityR :: Term -> Inference Theorem

Restricted combination rules

combineLeftR :: Term -> Theorem -> Inference Theorem

combineRightR :: Term -> Theorem -> Inference Theorem

Lambda-abstraction

abstractsR :: [(String, Type)] -> Theorem -> Inference Theorem

Generalised for of abstract, performing many abstractions one after the other.