Mosquito.Kernel.QualifiedName

Description

A module implementing qualified names of the form Mosquito.Bool.true and similar used for identifying constants, types, lemmas, and related notions throughout the Mosquito code base.

Qualified names and components

type Path = [String]

The type of paths (i.e. the . separated list of identifiers preceeding the ultimate dot in a qualified name). Can possibly be empty for top-level unqualified names like true.

data QualifiedName

The type of qualified names, consisting of a possibly empty path followed by a mandatory identifier (the head.)

Utility functions and tests

qualifiedNamePath :: QualifiedName -> Path

Returns the path of a qualified name.

qualifiedNameHead :: QualifiedName -> String

Returns the head of a qualified name.

mkQualifiedName :: Path -> String -> QualifiedName

Creates a new qualified name from a path and head. XXX: shouldn't this fail if the head is the empty string, or should those checks be elsewhere, e.g. in the parser?

mkSimpleName :: String -> QualifiedName

Constructs a `simple name', that is a qualified name whose path is empty and consists solely of a head.

Fresh name generations

freshQualifiedName

Arguments

:: Path
The path to put the name in

-> Maybe String
A suggested base for the fresh name

-> Set String
A set of names to avoid

-> QualifiedName

Creates a freshly generated qualified name in a given path.

Utilities for pretty printing

isInfix :: QualifiedName -> Bool

Tests whether a qualified name is an infix name for pretty-printing purposes. XXX: this is a really bad test and is broken. Fixme.

partialInfixKernel :: QualifiedName -> String

Recovers the infix kernel of a qualified name deemed to be infix. For instance, in the qualified name Mosquito.Bool._∧_ the kernel is ∧.