Mosquito.Kernel.Term

Description

The Mosquito kernel, defining types, terms, theorems, and providing primitive inference rules upon which everything else is built, along with simple mechanisms for extending the Mosquito logic (e.g. declaring axioms, defining new types and constants).

Everything in this file is considered trusted, that is, to trust the logical soundness and consistency of the whole system you must trust that this file correctly implements a version of classical, extensional higher-order logic, a la HOL4 and Isabelle/HOL. Due to the internal representation of terms, types and theorems not being exposed to the rest of the system (via Haskell's abstraction mechanisms), and with terms, types and theorems being correct by contruction (if this file is bug free), consistency of the whole system is guaranteed (the so-called LCF approach).

Some useful definitions

type Arity = Int

Type representing the arity of type operators and constants.

The inference monad

data Inference a

The inference monad, an error monad used throughout Mosquito to signal success or failure of a computation.

fail :: String -> Inference a

A function signifying a failing computation within the Inference monad. Parameter is the error message that will be displayed when the error is encountered at top-level.

inference :: Inference a -> (String -> b) -> (a -> b) -> b

An elimination principle for the Inference monad.

userMark :: [String] -> Inference ()

Records the invocation of an inference rule outside the kernel. When a later call in the Inference monad fails, the debug information provided as an argument is printed to the screen in a diagnostic message. By convention, the first element in the supplied argument list is the name of the inference rule being invoked.

Type operator descriptions

data TypeOperatorDescription

The TypeOperatorDescription type describes HOL type operators. Type operators are either primitive (in the case of the type of propositions and the function space arrow) or are defined. Defined type operators carry a definition around with them.

isDefinedTypeOperatorDescription :: TypeOperatorDescription -> Bool

Tests whether the description refers to a defined, rather than a primitive, type operator description.

isPrimitiveTypeOperatorDescription :: TypeOperatorDescription -> Bool

Tests whether the description refers to a primitive, rather than defined, type operator description.

typeOperatorDescriptionArity :: TypeOperatorDescription -> Arity

Retrieves the arity of a type operator description.

typeOperatorDescriptionQualifiedName :: TypeOperatorDescription -> QualifiedName

Retrieves the name of a type operator description.

typeOperatorDescriptionDefinition :: TypeOperatorDescription -> Maybe Theorem

Retrieves the definition of a defined type operator description. If the description is a primitive type operator, then the function returns Nothing.

boolQualifiedName :: QualifiedName

The name of the primitive HOL bool type.

functionQualifiedName :: QualifiedName

The name of the primitive HOL function space type operator.

boolDescription :: TypeOperatorDescription

The description of the primitive HOL bool type.

functionDescription :: TypeOperatorDescription

The description of the primitive HOL function space type former.

HOL types

data Type

The type of HOL types. Types are either type variables, or are operators which when applied to a list of other types, the length of which must match the arity proscribed, form another type.

data TypeView

A view type for HOL types, for pattern matching purposes outside of the kernel.

Constructors

TyVarView String

TyOperatorView TypeOperatorDescription [Type]

typeView :: Type -> TypeView

A view function that converts HOL types into their views, for pattern matching purposes outside of the kernel.

isTyVar :: Type -> Bool

Tests whether a Type is a type variable.

isTyOperator :: Type -> Bool

Tests whether a Type is a type operator.

isFunction :: Type -> Bool

Tests whether a Type is a function type.

isProposition :: Type -> Bool

Tests whether a Type is a propositional (Boolean) type.

fromTyOperator :: Type -> Inference (TypeOperatorDescription, [Type])

Deconstructs a type operator into its TypeOperatorDescription and [Type] components, failing with an error message if the input Type is not a type operator otherwise.

fromTyVar :: Type -> Inference String

Deconstructs a type variable into its String component, failing with an error message if the input Type is not a type variable otherwise.

fromFunction :: Type -> Inference (Type, Type)

Deconstructs a function type into its domain and range types, failing with an error message if the input Type is not a function type otherwise.

mkTyVar :: String -> Type

Makes a type variable from a String.

mkTyOperator :: TypeOperatorDescription -> [Type] -> Inference Type

Makes a type operator from a TypeOperatorDescription and a list of Type. Fails with an error message if the input list length differs from the declared arity of the TypeOperatorDescription.

alphaName :: String

Utility: the name of the alpha type variable, used quite often (but entirely arbitrarily) throghout Mosquito.

alphaType :: Type

The type variable corresponding to the alphaName above.

boolType :: Type

The Boolean type.

mkFunctionType :: Type -> Type -> Type

Utility function for constructing a function type from two previously defined types.

tv :: Type -> Set String

Collects the free type variables of a type (by definition, as we have no binding of type variables in types in Mosquito, all occurrences of type variables in a type are deemed free) into a Set.

Constant descriptions

data ConstantDescription

The description of constants defined within the Mosquito logic. Mosquito constants fall into one of two types: primitive, such as equality, or defined, which are defined by the user (following suitable checks to ensure consistency remains). All constants have a qualified name and type. Defined constants also keep hold of the defining term used in their definition to ensure that inconsistencies caused by mixing and matching constants of the same name but defined in different proof assistant states cannot creep into the system.

isDefinedConstantDescription :: ConstantDescription -> Bool

Tests whether a ConstantDescription is defined.

isPrimitiveConstantDescription :: ConstantDescription -> Bool

Tests whether a ConstantDescription is primitive.

isTypeRepresentationConstantDescription :: ConstantDescription -> Bool

Tests whether a ConstantDescription is a type representation constant, used in the declaration of a new type.

isTypeAbstractionConstantDescription :: ConstantDescription -> Bool

Tests whether a ConstantDescription is a type abstraction constant, used in the declaration of a new type.

constantDescriptionType :: ConstantDescription -> Type

Returns the type of a ConstantDescription.

constantDescriptionQualifiedName :: ConstantDescription -> QualifiedName

Returns the qualified name of a ConstantDescription.

constantDescriptionDefinition :: ConstantDescription -> Maybe Term

Returns the definition of a defined ConstantDescription. Fails if the ConstantDescription is primitive.

equalityType :: Type

Polymorphic type of the equality constant.

equalityQualifiedName :: QualifiedName

Qualified name of the (primitive) equality constant baked into Mosquito's higher-order logic.

equalityDescription :: ConstantDescription

ConstantDescription describing Mosquito's equality constant.

HOL terms

data Term

The type of HOL terms, explicitly-typed Church-style terms of the lambda-calculus extended with constants.

data TermView

A view type for HOL terms which allows pattern matching outside of the kernel.

Constructors

VarView String Type

ConstView ConstantDescription

AppView Term Term

LamView String Type Term

termView :: Term -> TermView

A view function for converting HOL terms into their views.

mkVar :: String -> Type -> Term

Makes a variable.

mkConst :: ConstantDescription -> Term

Makes a constant.

mkApp :: Term -> Term -> Inference Term

Makes an application. Fails if there is a type mismatch between the two arguments.

mkLam :: String -> Type -> Term -> Term

Makes a lambda-abstraction.

isVar :: Term -> Bool

Tests whether a Term is a variable.

isConst :: Term -> Bool

Tests whether a Term is a constant.

isApp :: Term -> Bool

Tests whether a Term is an application.

isLam :: Term -> Bool

Tests whether a Term is a lambda-abstraction.

fromVar :: Term -> Inference (String, Type)

Deconstructs a term, succeeding if the term is a variable, failing otherwise.

fromConst :: Term -> Inference ConstantDescription

Deconstructs a term, succeeding if the term is a constant, failing otherwise.

fromApp :: Term -> Inference (Term, Term)

Deconstructs a term, succeeding if the term is an application, failing otherwise.

fromLam :: Term -> Inference (String, Type, Term)

Deconstructs a term, succeeding if the term is a lambda-abstraction, failing otherwise.

Type checking

typeOf :: Term -> Inference Type

Computes the type of a term.

Alpha-equivalence and free variables

fv :: Term -> Set String

Collects the free variables (i.e. variables appearing within a term not bound by a lambda-abstraction) into a Set.

fvs :: (Foldable f, Functor f) => f Term -> Set String

Collects the free variables of a list of terms into a Set.

permute :: String -> String -> Term -> Term

Performs a swapping of names within terms. Used to define alpha-equivalence later.

Equality within the logic

equality :: Term

The equality constant.

isEquality :: Term -> Bool

Tests whether a term is an equality.

fromEquality :: Term -> Inference (Term, Term)

Deconstructs a term, succeeding if the term was an equality, failing otherwise.

mkEquality :: Term -> Term -> Inference Term

Makes an equality. Fails if there is a type mismatch between the two arguments and the equality constant.

Substitutions

data Substitution a

Generic substitutions (i.e. finite mappings from Strings to some other type).

empty :: Substitution a

The empty substitution.

compose :: Substitution a -> Substitution a -> Substitution a

Composes two substitutions.

mkSubstitution :: [(String, a)] -> Substitution a

Makes a substitution from an association list.

Type substitutions

typeSubst :: Substitution Type -> Type -> Type

Perform a type substitution replacing type variables whose names match the first argument with the second argument.

Term substitutions

termSubst :: Substitution Term -> Term -> Term

Performs a capture-avoiding term substitution with fresh-name generation if necessary.

termTypeSubst :: Substitution Type -> Term -> Term

Performs a type substitution on all types decorating the term (at lambda binding sites, within constant declarations and on decorating types appearing on variables).

HOL theorems

data Provenance

A provenance tag for theorems. Indelibly marks axioms with a FromAxiom tag. Everything that is not introduced as an axiom via the primitiveNewAxiom function is said to be derived safely.

track :: Provenance -> Provenance -> Provenance

Takes two provenance tags and produces a new one. If either of the tags provided as input are FromAxiom then the output will also be FromAxiom, otherwise will be DerivedSafely. Used to update provenance information in the primitive inference rules of the Mosquito logic provided below.

data Theorem

A theorem consists of a provenance tag detailing how it was derived (or more accurately, if any axiom was ever used in the derivation) combined with a list of hypotheses and a single conclusion. All terms making up the hypotheses and conclusion must be boolean-typed.

hypotheses :: Theorem -> [Term]

Obtain the hypotheses of a theorem.

conclusion :: Theorem -> Term

Obtain the conclusion of a theorem.

provenance :: Theorem -> Provenance

Obtain the provenance flag of a theorem.

union :: [Term] -> [Term] -> [Term]

Alpha-equivalent aware set-like union of term lists.

delete :: Term -> [Term] -> [Term]

Alpha-equivalent aware set-like delete of term lists.

deleteTheorem :: Term -> [Theorem] -> [Theorem]

Basic HOL theorems

alphaR :: Term -> Term -> Inference Theorem

Produces a derivation of {} |- t = t' given two terms t and t' that are alpha-equivalent.

symmetryR :: Theorem -> Inference Theorem

transitivityR :: Theorem -> Theorem -> Inference Theorem

Produces a derivation of Gamma u Delta |- t = v given a derivation of Gamma |- t = s and Delta |- s = u for all t, u and v.

abstractR :: String -> Type -> Theorem -> Inference Theorem

Produces a derivation of Gamma |- fn x:ty. t = fn x:ty. u given a derivation of the form Gamma |- t = u providing x does not appear free in the context Gamma.

combineR :: Theorem -> Theorem -> Inference Theorem

Produces a derivation of Gamma u Delta |- f x = g y given two derivations of the form Gamma |- f = g and Delta |- x = y.

etaR :: Term -> Inference Theorem

Produces a derivation of {} |- fn x:ty. (t x) = t when x is not in the free variables of t. Note that unlike HOL-light we take this as a primitive inference rule in the kernel, as opposed to taking it as an axiom later.

betaR :: Term -> Inference Theorem

Produces a derivation {} |- (fn x:ty. t)u = t[x := u] given an application. Note that this derivation rule is stronger than its HOL-light counterpart, as we permit full beta-equivalence in the kernel via this rule.

assumeR :: Term -> Inference Theorem

Produces a derivation {p} |- p for p a term of type Bool.

equalityModusPonensR :: Theorem -> Theorem -> Inference Theorem

Produces a derivation of Gamma u Delta |- q given two derivations of Gamma |- p = q and Delta |- p respectively.

deductAntiSymmetricR :: Theorem -> Theorem -> Inference Theorem

Produces a derivation of (Gamma - q) u (Delta - p) |- p = q from a pair of derivations of Gamma |- p and Delta |- q.

typeInstantiationR :: Substitution Type -> Theorem -> Inference Theorem

instantiationR :: Substitution Term -> Theorem -> Inference Theorem

Extending the logic

primitiveNewDefinedConstant :: QualifiedName -> Term -> Type -> Inference (Term, Theorem)

Produces a new defined constant. The right hand side of the definition must be closed (i.e. no free variables), and the free type variables of the right hand side must be a subset of the free type variables of the constant's type. Returns the new constant and a defining theorem.

primitiveNewDefinedType :: QualifiedName -> Theorem -> Inference (Theorem, Theorem, Term, Term, TypeOperatorDescription)

Defines a new inhabited type in provable bijection with a subset of an existing type.

primitiveNewAxiom :: Term -> Inference Theorem

Introduces an axiom. Dangerous!