Mosquito.ProofState.Conversionals

Description

Conversions, possibly failing mappings from terms t to theorems |- t = u and conversionals, higher-order functions over conversions. Conversions and conversionals can be used to implement a simple form of rewriting.

The conversion type, and simple conversions

type Conversion = Term -> Inference Theorem

failConv :: Conversion

The conversion that always fails.

successConv :: Conversion

The conversion that always succeeds (reflexivity).

conversionR :: Conversion -> Theorem -> Inference Theorem

Converts a conversion into a rule.

Conversionals

thenConv :: Conversion -> Conversion -> Conversion

The sequencing conversional.

repeatConv :: Conversion -> Conversion

Repetition of conversion application.

elseConv :: Conversion -> Conversion -> Conversion

The choice conversional. If the first conversion fails to apply then the second conversion is applied.

tryConv :: Conversion -> Conversion

Exception handling conversional. If the first conversion fails to apply, then the conversion that always succeeds is applied.

Syntax specific conversionals

combineConv :: Conversion -> Conversion -> Conversion

Apply a conversion to the left hand side of a combination, and another to the right hand side.

combineLConv :: Conversion -> Conversion

Specialisation of combineConv which applies a conversion to the left hand side of a combination, only.

combineRConv :: Conversion -> Conversion

Specialisation of combineConv which applies a conversion to the right hand side of a combination, only.

abstractConv :: Conversion -> Conversion

Applies a conversion to the body of an abstraction, under the binder.

More refined conversionals

unfoldConstantConv :: Theorem -> Conversion

Unfolds a constant with the constant's defining theorem.

replaceEverywhereConv :: Conversion -> Conversion

Applies a conversion at every point in a term, rewriting at all positions.

data Path

Paths for navigating through terms.

Constructors

GoLeft
Go left through an application.

GoRight
Go right through an application.

GoUnder
Go under a lambda abstraction.