Mosquito.ProofState.PreTactics

Description

Implements some basic tactics corresponding to the rules found in the kernel and in the file DerivedRules.hs.

Documentation

type Justification = [Theorem] -> Inference Theorem

type LocalEdit = [Term] -> Term -> Inference (Justification, [([Term], Term)])

type TermLocalEdit = Term -> LocalEdit

type TheoremLocalEdit = Theorem -> LocalEdit

data PreTactic

type TermPreTactic = Term -> PreTactic

type TheoremPreTactic = Theorem -> PreTactic

preTacticName :: forall arr. Arrow arr => Lens arr PreTactic String

localEdit :: forall arr. Arrow arr => Lens arr PreTactic LocalEdit

mkPreTactic :: String -> LocalEdit -> PreTactic

alphaL :: LocalEdit

alphaP :: PreTactic

symmetryL :: LocalEdit

symmetryP :: PreTactic

abstractL :: LocalEdit

abstractP :: PreTactic

combineL :: LocalEdit

combineP :: PreTactic

betaL :: LocalEdit

betaP :: PreTactic

etaL :: LocalEdit

etaP :: PreTactic

equalityModusPonensL :: TermLocalEdit

equalityModusPonensP :: TermPreTactic

assumeL :: LocalEdit

assumeP :: PreTactic

solveL :: TheoremLocalEdit

solveP :: TheoremPreTactic

conversionL :: Conversion -> LocalEdit

conversionP :: Conversion -> PreTactic

unfoldConstantP :: Theorem -> PreTactic

betaReduceP :: PreTactic

etaReduceP :: PreTactic