Mosquito.ProofState.ProofState

Description

The Mosquito proof state for tracking the current state during a backwards (tactic-oriented) proof. Functions for making an initial conjecture, applying pretactics to a proof state, and for retrieving a theorem from a complete backwards proof.

Documentation

data Selection where

Constructors

Selected :: Selection

Unselected :: Selection

data IncompleteDerivation

data ProofState

mkConjecture :: String -> Term -> Inference ProofState

mkConjectureRule :: String -> [Term] -> Term -> Inference ProofState

qed :: ProofState -> Inference Theorem

act :: ProofState -> Tactic -> Inference ProofState