Mosquito.ProofState.Unfolding

Documentation

unfoldTactic :: TheoremTactic

Unfolds a definition supplied as a theorem and then immediately solves extraneous subgoals, changing the goal to prove into the original goal with the constant unfolded.

pointwiseTactic :: Tactic

Decompose an equality between terms into equalities between subterms, solving easy goals with automation.

reductionLocalEdit :: LocalEdit

Performs a beta-reduction of the goal, immediately closing extraneous subgoals with automation.

reductionPreTactic :: PreTactic

reductionTactic :: Tactic