Mosquito.TermUtilities

Description

Provides some utilities for working with data types provided in the kernel that can safely be moved outside of the kernel.

Utility functions

partialFromSuccess :: Inference a -> a

Partial function extracting the result of a successful computation in the Inference monad, throwing an exception if the computation had not been successful.

freshs :: Int -> Maybe String -> Set String -> [String]

Generates a list of fresh names of a given length, using a suggested base name (otherwise defaulting to f as the base) and avoiding any names present in the provided set of names.

Useful deconstruction functions

fromBinaryOperation :: Term -> Inference (Term, Term, Term)

Deconstructs an arbitrary binary operation. The operation is the first element return, the left and right arguments of the operation follow.

Useful construction funtions

mkVars :: [(String, Type)] -> Inference [Term]

Generalised form of mkVar. Creates a list of variables based on a given association list between names of variables and types.

mkLams :: [(String, Type)] -> Term -> Term

Generalised form of mkLam. Creates an n-fold lambda-abstraction over a given term based on a given association list of names of variables to abstract and their types.

Type utilities

unifyTypes :: Type -> Type -> Inference (Substitution Type)

Unifies two types.