Mosquito.Utility.Pretty

Description

A utility module used for pretty-printing values throughout the Mosquito code base.

Size metrics and pretty printing

class Size a where

A typeclass capturing the notion of values that can be given a size for pretty-printing purposes. The size of a value should bear some relationship to the actual size of the string representation of that value under pretty-printing, for best results.

Methods

size :: a -> Int

sizes :: (Foldable f, Functor f, Size a) => f a -> Int

Sums the sizes of a series of metricisable values stored in a Foldable container.

class Pretty a where

A typeclass capturing the notion of values that can be pretty-printed. This is distinct from the Haskell Prelude Show typeclass, which is reserved for producing a faithful parsable (by GHCi) representation of a value.

Methods

pretty :: a -> String

Utility functions

bracket :: (Pretty a, Size a) => a -> String

Brackets a pretty-printed value with parentheses if its size, as calculated via the Size typeclass, is over a given threshold, otherwise returns the unmolested output of the pretty function.

lift :: Pretty a => (String -> b) -> a -> b

Lifts a function taking a String to an arbitrary output to a function working over pretty-printable input.

Pretty-printed I/O

putStr :: Pretty a => a -> IO ()

Version of putStr which first transforms its input into a pretty-printed representation with pretty before performing the IO action.

putStrLn :: Pretty a => a -> IO ()

Version of putStrLn which first transforms its input into a pretty-printed representation with pretty before performing the IO action.